requirements.txt
ipykernel
datasets
python-dotenv
scipy
matplotlib
transformers
ipywidgets
scikit-learn
accelerate
wandb
huggingface_hub
seaborn
swifter
nltk

bitsandbytes
dill
evaluate
fsspec
hf-transfer
lexical-diversity
multiprocess
sympy

trl
